- Editing files for proper OS X support.
- Fixing build warnings brought to light during build.

+++ src/ResourceLoader.cpp
@@ -529,7 +529,7 @@ void ResourceLoader::RemoveImage(Image *pImage)
 {
     SDL_SemWait(pQueueSemaphore);
 
-    for (int i = smartSpriteQueue.size() - 1; i >= 0; i--)
+    for (int i = static_cast<int>(smartSpriteQueue.size()) - 1; i >= 0; i--)
     {
         if (smartSpriteQueue[i] == pImage)
         {
